tdf#127372: PPTX: Shape's background transparency changes during RT
Need to use a better null value for FillTransparenceGradient.
Use black-black which has equal effect than having no transparency at all.

XFillFloatTransparenceItem has also an enabled attribute, but it
can not be accessed via the UNO API. We can acces only the gradient object,
so with a better null value we can check whether the API attribute is
actually set or not.

Change-Id: I480489f588376582f80d5062145b7f398ef61ad7
Reviewed-on: https://gerrit.libreoffice.org/78657
Reviewed-by: Tamás Zolnai <tamas.zolnai@collabora.com>
Tested-by: Tamás Zolnai <tamas.zolnai@collabora.com>