tdf#109158 improve load time of autocorrect XML file
Before, it seems to be doing a lot of unnecessary work because, by default,
startFastElement will call startUnknownElement, which constructs various objects, etc.

It seems to me that if the class already implements createFastChildContext,
it is already using XFastParser, and so it should implement startFastElement to avoid that work.

By avoiding going down the startUnknownElement path, reduces load time from
1.7s to 1.4s

Change-Id: Ifd745cf2cc22a8cb988a5668332cf6e70bb996da
Reviewed-on: https://gerrit.libreoffice.org/79269
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Noel Grandin <noel.grandin@collabora.co.uk>