tdf#124600 sw AddVerticalFrameOffsets: fix bad wrap of half-intersecting frame
Commit d07fc485d46f431405a3f6a002f951a08c559677 (tdf#116486 Consider
upper margin in paragraph positioning with flys, 2018-04-10) already
improved the interaction between paragraph upper margin and fly frames
in Word compat mode, this improves the situation further.

The problem is that in case the paragraph with an upper margin
intersects with a fly frame, then it makes sense to take the area of the
upper margin into account (i.e. intersect the fly frame rect with the
frame area, not with the smaller print area), but in case the
intersection covers only the margin area and none of the print area,
then the intersection is ignored by Word, so we should do the same in
compat mode. [ The "intersection bottom is above the para print area
top" condition can be true only in compat mode. ]

In other words, the mentioned commit fixed the position of a line of
text below a fly frame, this one makes sure that a long line is not
wrapped into 2 lines (referred to "text is covered" in the bugreport) in
Writer, to match Word's behavior.

Change-Id: I4581021c3dd570fbc8a89002394a43ef18f11a67
Reviewed-on: https://gerrit.libreoffice.org/79431
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins