tdf#98603 export runs with correct lang attribute (2/2).
Obtain lang attribute of rPr element from CharLocale,
CharLocaleAsian, or CharLocaleComplex based on the script
type of exported text.

There are several other call sites of WriteRunProperties,
like those in chartexport, and those for text fields,
Here I leave the default to CharLocale so it fallback
to western locale as it did before.

Change-Id: I33679517dd88f9e415933decc02f10afc807ad3b
Reviewed-on: https://gerrit.libreoffice.org/79973
Tested-by: Jenkins
Reviewed-by: Mark Hung <marklh9@gmail.com>