tdf#127785 compensate 180deg rotation for position of text box
If a shape is vertically flipped, then the text is rotated by 180deg
around the shape center. The rotation is done by SdrTextObj. There
text rectangle and shape rectangle are the same, so the rotated text
rectangle has the correct position in regard to flipping. But the
text rectangle for SdrObjCustomShape is set by the shape author in
the TextFrames attribute and might have an asymmetric position. The
patch translates the text rectangle so, that is will be at the
correct position after the 180deg rotation.

Change-Id: I42e552394cc4b0103530eab48750dd621c52cc5e
Reviewed-on: https://gerrit.libreoffice.org/79984
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>