Simplify a bit SfxAllEnumItem (svl)
Make private GetValueByPos which is only used in svl/source/items/aeitem.cxx
Remove "RemoveValue" method which is only used in InsertValue( sal_uInt16 nValue, const OUString &rValue )
It allows to call once "GetPosByValue" and the assert is useless since we're in the case:
"else if ( nPos != USHRT_MAX )"

I think we can more optimize by replacing vector by another container for pValues (type "SfxAllEnumValueArr").
Indeed, we insert and remove a lot in pValues in a specific position. Vector is not the best container for this
There's also still the part "//FIXME: Optimisation: use binary search or SortArray" in GetPosByValue_ method

Change-Id: I0cafea70e9d0361ef9a7b345ff770a9b888ef85b
Reviewed-on: https://gerrit.libreoffice.org/80089
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins
Reviewed-by: Julien Nabet <serval2412@yahoo.fr>