DOCX import: fix interaction of table and paragraph style in table cells
Both table and paragraph styles can define paragraph properties for the
content of table cells, e.g. line spacing. The intended behavior is that
direct formatting has priority, then paragraph style, then the table
styles.

But in practice table style had priority: table style is turned into a
set of property name-value pairs by writerfilter/ code, then this is
applied to the content of a cell using SwXCell::setPropertyValue(). To
make sure that direct-format-from-table-style doesn't override
direct-format, there was already a check to not replace properties which
are set directly.

The problem is that in case the property value comes from a paragraph
style, that's not direct and still should have priority over
direct-format-from-table-style.

Fix this by checking for custom values not only in the paragraph's item
set, but also in its parents. Unless the parent would be the default
style, which doesn't count. Or in case the paragraph is numbered, which
complicates the situation, so leave that case unchanged.

Change-Id: Ib554247cdc51fee0d9a6c7a26aecd38442dfa692
Reviewed-on: https://gerrit.libreoffice.org/80146
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins