related tdf#99602 ww8/rtfoutput: fix incorrect rounding on subscripts
Adding .5 is a poor mans version of rounding which works fine with
unsigned numbers, but not with negative numbers. Perhaps use the
exotic round() function instead?

In addition, the font size isn't necessarily an integer,
so that should have been a float.

The result of bad rounding was losing a percentage of the subscript
every round-trip.

Change-Id: I9093f7bfcd0b87249b42562668e45480dcb59f53
Reviewed-on: https://gerrit.libreoffice.org/80218
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>