tdf#127785 Apply 180deg compensation for flipV to text range
If a shape is vertically flipped, then the text is rotated by
180deg around the shape center. The rotation is done by SdrTextObj,
where text rectangle and shape rectangle are the same, so the
rotated text rectangle has the correct position despite flipping.
But the text rectangle for SdrObjCustomShape is set by the shape
author in the TextFrames attribute and might have an asymmetric
position. The patch compensates the flip-rotation by pre-rotate the
text rectangle.
This replaces commit caaa8fe7c4bb88185b5b11591ee8a619cff0eced.
The error in the old patch was, that it has uses a translation
instead of a rotation, and has used a wrong place. The result was,
that a text box, which has an own TextRotateAngle, had got a wrong
position.

Change-Id: Id38e8c1839afa5091cd251fc5237315ba7944263
Reviewed-on: https://gerrit.libreoffice.org/80310
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>
Tested-by: Regina Henschel <rb.henschel@t-online.de>
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>