Add some doc and fix comment in eppt (sd)
See https://docs.microsoft.com/en-us/openspecs/office_file_formats/ms-ppt/6be79dde-33c1-4c1b-8ccc-4b2301c08662
"
2.6.1:
A - fAutoAdvance (1 bit): A bit that specifies whether to automatically advance to the next slide
during a slide show based on timing information on the slide.
B - fWillSkipBuilds (1 bit): A bit that specifies whether to display animations during a slide show.
C - fUseSlideRange (1 bit): A bit that specifies whether to display only the slide range specified by
the startSlide and endSlide fields during a slide show.
D - fDocUseNamedShow (1 bit): A bit that specifies whether the slides shown during a slide show
are from the named show identified by namedShow. It MUST be ignored if fUseSlideRange is
TRUE.
E - fBrowseMode (1 bit): A bit that specifies whether the slide show is presented in a way optimized
for browsing. If fBrowseMode is TRUE, fKioskMode MUST be FALSE.
F - fKioskMode (1 bit): A bit that specifies whether the slide show is presented in a way optimized
to run at a kiosk. If fKioskMode is TRUE, fBrowseMode MUST be FALSE.
G - fWillSkipNarration (1 bit): A bit that specifies whether to play slide audio narrations during a
slide show.
H - fLoopContinuously (1 bit): A bit that specifies whether to restart the slide show at the
beginning after advancing from the last slide.
I - fHideScrollBar (1 bit): A bit that specifies whether to display the navigational scroll bar during a
slide show.
"
So yes, "IsEndless" corresponds to "fLoopContinuously"

Document "Skip narration"
+ "show scrollbar" corresponds to Bit 8 (the contrary of "fHideScrollBar")
so since this the Bit 8 of nFlags isn't set, it's always to 0, so we always show scrollbar

Change-Id: Ifb4f62a9f52cdafac9f788f4d6d1deb0cf1b8efa
Reviewed-on: https://gerrit.libreoffice.org/80415
Reviewed-by: Julien Nabet <serval2412@yahoo.fr>
Tested-by: Julien Nabet <serval2412@yahoo.fr>