Resolves: tdf#128137 cursor jumping during edit, etc.
The mbNeedUpdate hackery to block changes caused by the
user was subverted by the async handling of the treeview selection.

a) add a call to wait and process outstanding async events within
the protected block

b) move the needupdate down to the range manager to block updates during
this block in CheckForFormulaString too

c) when updating the visible rows, also update the selected rows even if they
are not visible.

Change-Id: I4c67d6cd2780f55d5eab5f856f6fda21703b4977
Reviewed-on: https://gerrit.libreoffice.org/80794
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>