tdf#127908 tdf#128193 Chart OOXML: Fix Custom Y axis title position
Import part: set the anchor position to the TOP_LEFT corner of the rectangle
during the import, if the textbox is rotated with 90 or 270 degree. Because
the OOXML files always contains the TOP_LEFT coordinates of a textbox, even if
they are rotated.

Note: Unfortunatelly we do not know the shape size, so this fix
cannot handle rotations different than 0, 90 or 270 degrees.

Export part: export the top left corner coordinates of axis title shape
as the OOXML Standerd requires.

Change-Id: Id0875d65884f6bfef8726135a7c03418d2ce3f23
Reviewed-on: https://gerrit.libreoffice.org/80939
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>
Tested-by: László Németh <nemeth@numbertext.org>