tdf#42748 show changes of list numbers and bullets
At splitting or joining list items or adding new ones, the
removed or new list number or bullet didn't show the change,
only a vertical line added to the left margin of the list
items. Moreover, deleting only the text of the list
item (keeping its number) looked like the full deletion of
the item.

Change-Id: I4568713b5c94584761609d891b3eb36c45f8502c
Reviewed-on: https://gerrit.libreoffice.org/81258
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>