tdf#127412 sw: fix reported ZOrder of shape+textbox pairs
Writer documents may have shape+textbox pairs, in which case internally
they have individual ZOrders, but the textbox ZOrder is always just 1
larger than their shape, and externally this is not visible (UNO API,
UI).

This is implemented by e.g. SwXDrawPage::getCount(), which ignores those
textboxes. This worked in general for SwVirtFlyDrawObj, but in case the
textbox is a SwFlyDrawObj, then these were not filtered out. Fix this,
so the scenario when the shapes are added to the draw page following
ZOrder (0, 1, etc) works (internally producing 0, 1, 2, 3, etc ZOrders).

Change-Id: I2a04fb76029d83390d418c764fdfbe7a1ee0f208
Reviewed-on: https://gerrit.libreoffice.org/81277
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>