Related: tdf#125038 DOCX import: fix unexpected MERGEFIELD result inside IF
The problem is that DOCX supports nesting MERGEFIELD fields inside IF
fields, while SwHiddenTextField only supports a single string as a
condition.

This means in case there are MERGEFIELD fields inside the IF field,
those fields will be inserted to the doc model before the IF field,
exposing their value, while Word only uses their value during the
evaluation of the IF expression.

Fix the problem by inspecting the parent field command before setting
the MERGEFIELD result.

Change-Id: Ieca098f16f756bab5d23f219fa4ca30d077d4bb7
Reviewed-on: https://gerrit.libreoffice.org/81615
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins