Related: tdf#125038 DOCX import: fix unexpected linebreak inside IF condition
Writer body text is expected to only contain the result of the field. So
in case both the field command and the field result contains a
linebreak, we need to make sure that linebreaks are ignored in the field
command for field types where the Writer field implementation expects a
single string.

With this, the number of paragraphs in the bugdoc is now correct.

Change-Id: I42f208d6943750ba2e8f88b52c373f6ca9cb2b71
Reviewed-on: https://gerrit.libreoffice.org/81786
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins