tdf#38199 Readonly mode tab key: Don't scroll if no hyperlink found
When viewing a document in readonly mode the action of the tab key is to
move among Graphic, Frame, Ole, DrawObject, or DbForm if one these is
already selected. If the cursor is inside an input field and has no multi
selection then move among InputFields. Finally if neither the previous
two cases apply then move among Hyperlinks which causes cursor position
to move to top or bottom of the document when no Hyperlink is found.
When no Hyperlink is found this patch restores the cursor position to
the position before search and does not scroll to it on EndAction.

Change-Id: I94ac882e7be124d2ddb6f86daac03a9389117cc5
Reviewed-on: https://gerrit.libreoffice.org/81955
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>